split AT Functor into autovec and AVX512
